refactor(helpers): move common fn and vars to utils file
