+++ packages/_helpers/mixins.ts
@@ -1,20 +1,9 @@
-import { prop, ComponentProps, Component } from 'skatejs';
+import { prop, ComponentProps } from 'skatejs';
 import { ColorType } from './colorTypes';
 import { IS_DEV } from './environment';
-
-const { attachShadow } = HTMLElement.prototype;
-const { ShadyCSS } = window;
-const $template = Symbol();
-const nativeShadowDomSupport = attachShadow && attachShadow.toString().indexOf('native code') > -1;
+import { scopeCss, Constructable } from './utils';
 
 
-/** Any type that can construct *something*. */
-// export type Constructable<T> = new (...args: any[]) => T;
-export interface Constructable<T> {
-  readonly props: ComponentProps<any, any>;
-  readonly observedAttributes: string[];
-  new ( ...args: any[] ): T,
-}
 
 export type ColoredProps = {
   color?: ColorType,
@@ -63,11 +52,4 @@ export function Css<BC extends Constructable<{}>>( Base: BC) {
 
 }
 
-function scopeCss (elem: Component<any> & {[key: string]: any}, css: string): string | void {
-  if (nativeShadowDomSupport) {
-    return css;
-  }
-  const template = elem[$template] || (elem[$template] = document.createElement('template'));
-  template.innerHTML = `<style>${css}</style>`;
-  ShadyCSS.prepareTemplate(template, elem.localName);
-}
+
